// src/shared/generated/api/generated/api.ts
export interface OtaProject {
    /**
     * 
     * @type {string}
     * @memberof OtaProject
     */
    'project_id': string;
    /**
     * 
     * @type {string}
     * @memberof OtaProject
     */
    'file_path': string;
    /**
     * 
     * @type {string}
     * @memberof OtaProject
     */
    'owner_username': string;
    /**
     * 
     * @type {string}
     * @memberof OtaProject
     */
    'name': string;
    /**
     * 
     * @type {ProjectStatus}
     * @memberof OtaProject
     */
    'status': ProjectStatus;
    /**
     * 
     * @type {string}
     * @memberof OtaProject
     */
    'creation_date': string;
    /**
     * 
     * @type {string}
     * @memberof OtaProject
     */
    'updated_at': string;
    /**
     * 
     * @type {string}
     * @memberof OtaProject
     */
    'theme'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof OtaProject
     */
    'description'?: string | null;
    /**
     * 
     * @type {TranscriptionResult}
     * @memberof OtaProject
     */
    'transcription_result'?: TranscriptionResult | null;
    /**
     * 
     * @type {ProjectMetaInformation}
     * @memberof OtaProject
     */
    'meta_information'?: ProjectMetaInformation | null;
    /**
     * 
     * @type {string}
     * @memberof OtaProject
     */
    'resume_result'?: string | null;
    /**
     * 
     * @type {TranscriptionResult}
     * @memberof OtaProject
     */
    'editable_transcription_result'?: TranscriptionResult | null;
    /**
     * 
     * @type {ProjectLastOptions}
     * @memberof OtaProject
     */
    'last_options'?: ProjectLastOptions | null;
    /**
     * 
     * @type {ProjectStagesInformation}
     * @memberof OtaProject
     */
    'stages'?: ProjectStagesInformation | null;
    /**
     * 
     * @type {number}
     * @memberof OtaProject
     */
    'position_in_queue'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof OtaProject
     */
    'error_detail'?: string | null;
}
// ...
export interface ProjectLastOptions {
    /**
     * 
     * @type {string}
     * @memberof ProjectLastOptions
     */
    'force_language'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ProjectLastOptions
     */
    'batch_size'?: number | null;
    /**
     * Replace numerical digits with their pronunciation, increases diarization accuracy
     * @type {boolean}
     * @memberof ProjectLastOptions
     */
    'suppress_numerals'?: boolean;
    /**
     * Enable words extended info (It means: run wav2vec2 model after transcribation process to get accuracy and timings for every word)
     * @type {boolean}
     * @memberof ProjectLastOptions
     */
    'words_extended_information'?: boolean;
    /**
     * Noise reduction (DeepFilterNet for this task)
     * @type {boolean}
     * @memberof ProjectLastOptions
     */
    'noise_reduction'?: boolean;
    /**
     * Detect obscene words (only for RU language now)
     * @type {boolean}
     * @memberof ProjectLastOptions
     */
    'detect_obscene_words'?: boolean;
    /**
     * Threshold (upper bound) for obscene words. Don\'t change it if you don\'t know what it is.
     * @type {number}
     * @memberof ProjectLastOptions
     */
    'obscene_threshold'?: number;
    /**
     * 
     * @type {string}
     * @memberof ProjectLastOptions
     */
    'audio_type'?: ProjectLastOptionsAudioTypeEnum | null;
    /**
     * 
     * @type {number}
     * @memberof ProjectLastOptions
     */
    'num_speakers'?: number | null;
    /**
     * This tool separates track into vocals & non-vocals (cutting the noise)
     * @type {boolean}
     * @memberof ProjectLastOptions
     */
    'demucs_enabled'?: boolean;
    /**
     * Beautify speakers order (first speaker_id = 1, second speaker_id = 2, etc)
     * @type {boolean}
     * @memberof ProjectLastOptions
     */
    'beautify_speakers_order'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ProjectLastOptions
     */
    'is_diarization'?: boolean;
}

export const ProjectLastOptionsAudioTypeEnum = {
    Meeting: 'meeting',
    Telephonic: 'telephonic'
} as const;

export type ProjectLastOptionsAudioTypeEnum = typeof ProjectLastOptionsAudioTypeEnum[keyof typeof ProjectLastOptionsAudioTypeEnum];

/**
 * 
 * @export
 * @interface ProjectMetaInformation
 */
export interface ProjectMetaInformation {
    /**
     * 
     * @type {number}
     * @memberof ProjectMetaInformation
     */
    'file_duration_millis'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof ProjectMetaInformation
     */
    'file_size_bytes'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof ProjectMetaInformation
     */
    'task_duration_millis'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof ProjectMetaInformation
     */
    'words_count'?: number | null;
}
/**
 * 
 * @export
 * @interface ProjectStagesInformation
 */
export interface ProjectStagesInformation {
    /**
     * 
     * @type {boolean}
     * @memberof ProjectStagesInformation
     */
    'is_transcribed'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ProjectStagesInformation
     */
    'is_edited'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ProjectStagesInformation
     */
    'is_summarized'?: boolean;
}
/**
 * 
 * @export
 * @enum {string}
 */

export const ProjectStatus = {
    Ok: 'OK',
    InProgress: 'IN_PROGRESS',
    SummarizationGenerating: 'SUMMARIZATION_GENERATING',
    Error: 'ERROR'
} as const;

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];
// ...
export interface TranscriptionResult {
    /**
     * 
     * @type {string}
     * @memberof TranscriptionResult
     */
    'language_code': string;
    /**
     * 
     * @type {Array<TranscriptionSegment>}
     * @memberof TranscriptionResult
     */
    'sentences': Array<TranscriptionSegment>;
}
/**
 * 
 * @export
 * @interface TranscriptionSegment
 */
export interface TranscriptionSegment {
    /**
     * 
     * @type {number}
     * @memberof TranscriptionSegment
     */
    'start_millis': number;
    /**
     * 
     * @type {number}
     * @memberof TranscriptionSegment
     */
    'end_millis': number;
    /**
     * 
     * @type {string}
     * @memberof TranscriptionSegment
     */
    'text': string;
    /**
     * 
     * @type {number}
     * @memberof TranscriptionSegment
     */
    'speaker_id'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof TranscriptionSegment
     */
    'speaker_name'?: string | null;
    /**
     * 
     * @type {Array<WordExtendedInfo>}
     * @memberof TranscriptionSegment
     */
    'words_extended_information'?: Array<WordExtendedInfo> | null;
}
// ...
export interface WordExtendedInfo {
    /**
     * 
     * @type {string}
     * @memberof WordExtendedInfo
     */
    'word': string;
    /**
     * 
     * @type {number}
     * @memberof WordExtendedInfo
     */
    'start_millis': number;
    /**
     * 
     * @type {number}
     * @memberof WordExtendedInfo
     */
    'end_millis': number;
    /**
     * 
     * @type {number}
     * @memberof WordExtendedInfo
     */
    'accuracy': number;
    /**
     * 
     * @type {number}
     * @memberof WordExtendedInfo
     */
    'speaker_id'?: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof WordExtendedInfo
     */
    'is_obscene'?: boolean;
}